+++ src/main/java/nl/mindef/c2sc/nbs/olsr/pud/uplink/server/util/NmeaUtil.java
@@ -3,7 +3,7 @@ package nl.mindef.c2sc.nbs.olsr.pud.uplink.server.util;
 public class NmeaUtil {
 	public static double nmeaDeg2Ndeg(double degrees) {
 		long deg = (long) degrees;
-		double fra_part = degrees - (double) deg;
-		return (((double) deg * 100.0) + (fra_part * 60.0));
+		double fra_part = degrees - deg;
+		return ((deg * 100.0) + (fra_part * 60.0));
 	}
 }
